Fix implicit conversion error in event loop
Test: ./run_sim.sh

+++ core/event_loop.cc
@@ -339,7 +339,8 @@ bool EventLoop::allocateAndPostEvent(uint16_t eventType, void *eventData,
   Milliseconds receivedTime = Nanoseconds(SystemTime::getMonotonicTime());
   // The event loop should never contain more than 65 seconds worth of data
   // unless something has gone terribly wrong so use uint16_t to save space.
-  uint16_t receivedTimeMillis = receivedTime.getMilliseconds();
+  uint16_t receivedTimeMillis =
+      static_cast<uint16_t>(receivedTime.getMilliseconds());
 
   Event *event = mEventPool.allocate(eventType, receivedTimeMillis, eventData,
                                      freeCallback, senderInstanceId,
